requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
alibabacloud_dysmsapi20170525==2.0.24
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
attrs==25.3.0
Authlib==1.6.0
beautifulsoup4==4.8.2
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
compressed-rtf==1.0.7
crcmod==1.7
cryptography==42.0.8
cssselect2==0.7.0
dashscope==1.23.4
dnspython==2.6.1
docx2txt==0.9
ebcdic==1.1.1
email_validator==2.2.0
extract-msg==0.28.7
fastapi==0.111.0
fastapi-sessions==0.0.3
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
IMAPClient==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
lxml==5.4.0
MarkupSafe==3.0.2
multidict==6.4.4
numpy==1.26.4
olefile==0.47
opencv-python==4.9.0.80
oss2==2.19.1
packaging==25.0
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20191110
pillow==11.2.1
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
PyPDF2==3.0.1
pytesseract==0.3.10
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==0.6.23
redis==5.0.7
requests==2.32.4
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
starlette==0.46.2
textract==1.6.5
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.3
websocket-client==1.8.0
xlrd==1.2.0
XlsxWriter==3.2.3
yarl==1.20.1
